Use alternate method for searching vtables
They are now recofnized from rdata relocs where one has a value pointing
to rdata, followed immediately by one pointing to text. Should be faster
when looking for multiple sets of vtables as the rdata will only be fully
searched once, with the vtables cached and sorted for fast search.

The new vtable search was also made to support 64-bit.