Scarf update
Scarf fix ends up causing new dat warning (correctly) for not knowing
how to widen u8 write to global memory u8. Technically this specific
variable is u32, so it could be widened, but can't know that it's always
this one.

It's an ui global that stores previously shown status screen current tech
(Or depending on status screen unit also completely different things),
so samase should not care about it anyway.