Move primary dat analysis to new WidenInstruction modes too
Legacy mode can now be removed, the 32bit removed patches are widened
global accesses that shouldn't be widened under this logic, and were
(probably) confused as ebp-based reads. Technically the global is u32 so
it would be fine, but it is also a status screen global so it is not
important.

64bit additions are an improvement, as stack accesses through esp are now
correctly patched, but they do not get proper u32 alignment / extra
allocations like ebp-based accesses do, so some of the patches are wrong.