Add analysis for the save_replay function.
Newer versions take a lot less work because they added a unique string
reference, but for older versions I had to come at it from a layer up.

Versions 1.20.8 until 1.20.11 (exclusive) don't have the function
because it was inlined (honestly maybe they just pulled it out in
1.20.11 to create the crash replay dump thing). We could also grab the
calling function that creates a replay specifically in the replay dir,
although this feels a bit more annoying to use in general (especially
given the data directory weirdness we've seen on some machines), so I
just left those versions as None.